WIP on OverviewToolbar
